refactor: oauth/model에서 oauth/domain으로 패키지를 이동했습니다
