--- src/main/java/com/example/cherrydan/oauth/model/RefreshToken.java
@@ -1,28 +1,0 @@
-package com.example.cherrydan.oauth.model;
-
-import com.example.cherrydan.common.entity.BaseTimeEntity;
-import com.example.cherrydan.user.domain.User;
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Table(name = "auth_token")
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class RefreshToken extends BaseTimeEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(name = "refresh_token", nullable = false)
-    private String refreshToken;
-
-    @OneToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id", nullable = false)
-    private User user;
-}
