+++ src/main/java/com/example/cherrydan/oauth/security/oauth2/CustomOAuth2UserService.java
@@ -4,9 +4,9 @@ import com.example.cherrydan.common.exception.AuthException;
 import com.example.cherrydan.common.exception.ErrorMessage;
 import com.example.cherrydan.fcm.dto.FCMTokenRequest;
 import com.example.cherrydan.fcm.service.FCMTokenService;
+import com.example.cherrydan.oauth.domain.AuthProvider;
 import com.example.cherrydan.oauth.dto.AppleLoginRequest;
 import com.example.cherrydan.oauth.dto.LoginRequest;
-import com.example.cherrydan.oauth.model.AuthProvider;
 import com.example.cherrydan.oauth.security.jwt.UserDetailsImpl;
 import com.example.cherrydan.oauth.security.oauth2.exception.OAuth2AuthenticationProcessingException;
 import com.example.cherrydan.oauth.security.oauth2.user.OAuth2UserInfo;
